# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DashboardResponse": {
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "recent_reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemResponse"
            },
            "title": "Recent Reviews",
            "type": "array"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/ExecutionStatus"
            },
            "title": "Status Counts",
            "type": "object"
          },
          "total_reviews": {
            "title": "Total Reviews",
            "type": "integer"
          },
          "worker": {
            "$ref": "#/components/schemas/WorkerResponse"
          },
          "workers": {
            "items": {
              "$ref": "#/components/schemas/WorkerResponse"
            },
            "title": "Workers",
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "total_reviews",
          "status_counts",
          "worker",
          "workers",
          "recent_reviews",
          "next_cursor"
        ],
        "title": "DashboardResponse",
        "type": "object"
      },
...
      "ExecutionStatus": {
        "enum": [
          "queued",
          "ci",
          "planning",
          "agent_batches",
          "aggregating",
          "awaiting_approval",
          "approved",
          "rejected",
          "awaiting_publish",
          "publishing",
          "paused",
          "waiting_for_ci",
          "running",
          "ready_for_review",
          "completed",
          "failed",
          "timed_out",
          "cancelled",
          "superseded"
        ],
        "title": "ExecutionStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewAcceptedResponse": {
        "description": "\u5f02\u6b65\u5ba1\u67e5\u4efb\u52a1\u5165\u961f\u540e\u7684\u7a33\u5b9a\u786e\u8ba4\u54cd\u5e94\u3002",
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "execution_status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "review_run_id": {
            "title": "Review Run Id",
            "type": "string"
          },
          "review_task_id": {
            "title": "Review Task Id",
            "type": "string"
          },
          "review_version_key": {
            "title": "Review Version Key",
            "type": "string"
          }
        },
        "required": [
          "review_run_id",
          "review_task_id",
          "review_version_key",
          "execution_status",
          "accepted_at",
          "created"
        ],
        "title": "ReviewAcceptedResponse",
        "type": "object"
      },
...
      "ReviewItemResponse": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "base_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Ref"
          },
          "base_repository": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Repository"
          },
          "coverage_status": {
            "title": "Coverage Status",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "execution_status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "finding_count": {
            "title": "Finding Count",
            "type": "integer"
          },
          "head_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Ref"
          },
          "head_repository": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Repository"
          },
          "head_sha": {
            "title": "Head Sha",
            "type": "string"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_error_code": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Code"
          },
          "last_error_details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Details"
          },
          "last_error_retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Retryable"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "model_attempt_count": {
            "title": "Model Attempt Count",
            "type": "integer"
          },
          "model_review_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Review Completed At"
          },
          "pr_author_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Author Login"
          },
          "pr_html_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Html Url"
          },
          "pr_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Title"
          },
          "pull_request_number": {
            "title": "Pull Request Number",
            "type": "integer"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "review_conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Conclusion"
          },
          "review_run_id": {
            "title": "Review Run Id",
            "type": "string"
          },
          "review_task_id": {
            "title": "Review Task Id",
            "type": "string"
          },
          "unreviewed_finding_count": {
            "title": "Unreviewed Finding Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_status": {
            "$ref": "#/components/schemas/ExecutionStatus",
            "default": "queued"
          }
        },
        "required": [
          "review_run_id",
          "review_task_id",
          "repository",
          "pull_request_number",
          "head_sha",
          "pr_title",
          "pr_author_login",
          "pr_html_url",
          "head_repository",
          "head_ref",
          "base_repository",
          "base_ref",
          "execution_status",
          "attempt_count",
          "max_attempts",
          "last_error",
          "last_error_code",
          "last_error_retryable",
          "last_error_details",
          "review_conclusion",
          "coverage_status",
          "model_review_completed_at",
          "finding_count",
          "unreviewed_finding_count",
          "model_attempt_count",
          "created_at",
          "updated_at"
        ],
        "title": "ReviewItemResponse",
        "type": "object"
      },
      "ReviewListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items",
          "next_cursor"
        ],
        "title": "ReviewListResponse",
        "type": "object"
      },
...
      "ReviewRequest": {
        "additionalProperties": false,
        "description": "\u4e00\u6b21\u5f02\u6b65 Pull Request \u5ba1\u67e5\u4f7f\u7528\u7684\u5185\u90e8\u8bf7\u6c42\u3002",
        "properties": {
          "head_sha": {
            "maxLength": 64,
            "minLength": 40,
            "title": "Head Sha",
            "type": "string"
          },
          "installation_id": {
            "exclusiveMinimum": 0.0,
            "title": "Installation Id",
            "type": "integer"
          },
          "pull_request_number": {
            "exclusiveMinimum": 0.0,
            "title": "Pull Request Number",
            "type": "integer"
          },
          "repository": {
            "maxLength": 255,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$",
            "title": "Repository",
            "type": "string"
          },
          "repository_id": {
            "exclusiveMinimum": 0.0,
            "title": "Repository Id",
            "type": "integer"
          }
        },
        "required": [
          "installation_id",
          "repository_id",
          "repository",
          "pull_request_number",
          "head_sha"
        ],
        "title": "ReviewRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkerResponse": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "current_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Task Id"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "online": {
            "title": "Online",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkerStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "worker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Id"
          }
        },
        "required": [
          "configured",
          "online",
          "worker_id",
          "status",
          "current_task_id",
          "started_at",
          "last_seen_at"
        ],
        "title": "WorkerResponse",
        "type": "object"
      },
      "WorkerStatus": {
        "description": "\u961f\u5217 Worker \u8fdb\u7a0b\u4f7f\u7528\u7684\u7cbe\u7b80\u53ef\u89c2\u5bdf\u751f\u547d\u5468\u671f\u3002",
        "enum": [
          "starting",
          "idle",
          "busy",
          "stopping"
        ],
        "title": "WorkerStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "OpenReviewer API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboard": {
      "get": {
        "description": "\u8fd4\u56de\u8ba4\u8bc1\u540e\u7684\u4efb\u52a1\u7edf\u8ba1\u3001\u6700\u8fd1\u4efb\u52a1\u548c Worker \u72b6\u6001\u3002\n\n\u53c2\u6570\uff1a\n    _: \u4ec5\u7528\u4e8e\u89e6\u53d1\u7ba1\u7406\u5458\u4f1a\u8bdd\u6821\u9a8c\u7684\u4f9d\u8d56\u7ed3\u679c\u3002\n    limit: \u6700\u8fd1\u4efb\u52a1\u6570\u91cf\uff0c\u8303\u56f4 1 \u5230 100\uff0c\u9ed8\u8ba4 50\u3002\n\n\u8fd4\u56de\uff1a\n    \u5f53\u524d\u6570\u636e\u5e93\u5feb\u7167\uff1bWorker \u6ca1\u6709\u5fc3\u8df3\u65f6\u4f1a\u660e\u786e\u6807\u8bb0\u4e3a\u672a\u914d\u7f6e/\u79bb\u7ebf\uff0c\u800c\u4e0d\u662f\n    \u731c\u6d4b\u5065\u5eb7\u72b6\u6001\u3002\n\n\u5f02\u5e38\uff1a\n    HTTPException(401): \u4f1a\u8bdd\u65e0\u6548\u3002\n    HTTPException(503): Dashboard \u6570\u636e\u65e0\u6cd5\u8bfb\u53d6\u3002",
        "operationId": "get_dashboard_api_v1_dashboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard"
      }
    },
...
    "/api/v1/reviews": {
      "get": {
        "description": "\u8fd4\u56de\u8ba4\u8bc1\u540e\u7684\u6700\u8fd1\u5ba1\u67e5\u4efb\u52a1\u5217\u8868\u3002\n\n\u53c2\u6570\uff1a\n    _: \u7ba1\u7406\u5458\u4f1a\u8bdd\u4f9d\u8d56\u3002\n    limit: \u8fd4\u56de\u6761\u6570\uff0c\u8303\u56f4 1 \u5230 100\uff0c\u9ed8\u8ba4 50\u3002\n\n\u8fd4\u56de\uff1a\n    \u5305\u542b\u6570\u636e\u5e93\u4e2d\u7684\u603b\u8fd0\u884c\u6570\u548c\u6309\u521b\u5efa\u65f6\u95f4\u5012\u5e8f\u6392\u5217\u7684\u6700\u8fd1\u4efb\u52a1\uff1b\u4efb\u52a1\u72b6\u6001\u3001\n    \u5c1d\u8bd5\u6b21\u6570\u548c\u6700\u540e\u9519\u8bef\u6765\u81ea\u540c\u4e00\u6b21 Dashboard \u8bfb\u53d6\u3002\n\n\u5f02\u5e38\uff1a\n    HTTPException(401): \u4f1a\u8bdd\u65e0\u6548\u3002\n    HTTPException(503): \u67e5\u8be2\u5931\u8d25\u3002",
        "operationId": "list_reviews_api_v1_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reviews"
      },
      "post": {
        "description": "\u6821\u9a8c\u5e42\u7b49\u952e\u5e76\u63a5\u53d7\u4e00\u4e2a\u5f02\u6b65\u5ba1\u67e5\u4efb\u52a1\u3002\n\n\u8bf7\u6c42\u5fc5\u987b\u5148\u901a\u8fc7\u4f1a\u8bdd\u548c\u540c\u6e90\u68c0\u67e5\uff1b\u4ed3\u50a8\u5c42\u4fdd\u8bc1\u8fd0\u884c\u3001\u4efb\u52a1\u548c Outbox \u4e8b\u4ef6\u5728\u540c\n\u4e00\u4e2a\u4e8b\u52a1\u4e2d\u6301\u4e45\u5316\u3002\u76f8\u540c\u952e\u91cd\u590d\u63d0\u4ea4\u8fd4\u56de\u539f\u4efb\u52a1\uff0c\u4e0d\u540c\u5185\u5bb9\u590d\u7528\u540c\u4e00\u952e\u5219\u8fd4\u56de\n409\uff0c\u6570\u636e\u5e93\u6682\u65f6\u4e0d\u53ef\u7528\u5219\u8fd4\u56de\u53ef\u5b89\u5168\u91cd\u8bd5\u7684 503\u3002\n\n\u53c2\u6570\uff1a\n    request_body: \u5df2\u901a\u8fc7 Pydantic \u4e25\u683c\u5b57\u6bb5\u6821\u9a8c\u7684\u5ba1\u67e5\u8bf7\u6c42\u3002\n    idempotency_key: HTTP ``Idempotency-Key``\uff0c\u957f\u5ea6 1 \u5230 200\uff1b\u51fd\u6570\u4f1a\u518d\u53bb\u6389\n        \u9996\u5c3e\u7a7a\u767d\uff0c\u7a7a\u767d\u952e\u8fd4\u56de 422\u3002\n    _: \u7ba1\u7406\u5458\u4f1a\u8bdd\u4f9d\u8d56\u7ed3\u679c\uff0c\u4ec5\u7528\u4e8e\u786e\u8ba4\u8c03\u7528\u65b9\u5df2\u767b\u5f55\u3002\n    __: \u540c\u6e90\u4f9d\u8d56\u7ed3\u679c\uff0c\u4ec5\u7528\u4e8e\u963b\u6b62\u5e26\u6076\u610f Origin \u7684\u6d4f\u89c8\u5668\u526f\u4f5c\u7528\u8bf7\u6c42\u3002\n\n\u8fd4\u56de\uff1a\n    202 \u54cd\u5e94\uff0c\u5305\u542b\u8fd0\u884c/\u4efb\u52a1 ID\u3001\u7248\u672c\u952e\u3001\u5f53\u524d\u6267\u884c\u72b6\u6001\u3001\u9996\u6b21\u63a5\u53d7\u65f6\u95f4\u548c\n    ``created`` \u6807\u5fd7\u3002\u91cd\u590d\u8bf7\u6c42\u7684\u72b6\u6001\u53ef\u80fd\u5df2\u7ecf\u4ece ``queued`` \u63a8\u8fdb\u5230\u5176\u4ed6\u72b6\u6001\u3002\n\n\u5f02\u5e38\uff1a\n    HTTPException(401/403/422/409/503): \u5206\u522b\u5bf9\u5e94\u4f1a\u8bdd\u65e0\u6548\u3001Origin \u4e0d\u5339\u914d\u3001\n    \u8bf7\u6c42\u6216\u5e42\u7b49\u952e\u4e0d\u5408\u6cd5\u3001\u952e\u6307\u5411\u4e0d\u540c\u5185\u5bb9\u3001\u6301\u4e45\u5316\u4e0d\u53ef\u7528\u3002",
        "operationId": "create_review_api_v1_reviews_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review"
      }
    },
...
  }
}